+++ derex/runner/ddc.py
@@ -4,9 +4,9 @@ These wrappers invoke `docker-compose` functions to get their job done.
 They put a `docker.compose.yml` file in place based on user configuration.
 """
 from derex.runner.compose_utils import run_docker_compose
-from derex.runner.docker import check_services
-from derex.runner.docker import ensure_volumes_present
-from derex.runner.docker import is_docker_working
+from derex.runner.docker_utils import check_services
+from derex.runner.docker_utils import ensure_volumes_present
+from derex.runner.docker_utils import is_docker_working
 from derex.runner.logging_utils import setup_logging
 from derex.runner.plugins import get_sorted_items
 from derex.runner.plugins import setup_plugin_manager
