+++ tests/test_docker.py
@@ -10,8 +10,8 @@ MINIMAL_PROJ = Path(__file__).with_name("fixtures") / "minimal"
 
 
 def test_ensure_volumes_present(mocker):
-    from derex.runner.docker import ensure_volumes_present
-    from derex.runner.docker import VOLUMES
+    from derex.runner.docker_utils import ensure_volumes_present
+    from derex.runner.docker_utils import VOLUMES
 
     client = mocker.patch("derex.runner.docker.client")
 
@@ -28,7 +28,7 @@ def test_ensure_volumes_present(mocker):
 
 
 def test_check_services(mocker):
-    from derex.runner.docker import check_services
+    from derex.runner.docker_utils import check_services
 
     client = mocker.patch("derex.runner.docker.client")
 
@@ -42,7 +42,7 @@ def test_check_services(mocker):
 
 
 def test_wait_for_service(mocker):
-    from derex.runner.docker import wait_for_service
+    from derex.runner.docker_utils import wait_for_service
 
     container = mocker.MagicMock()
     container.exec_run.return_value = mocker.MagicMock(exit_code=0)
@@ -56,7 +56,7 @@ def test_wait_for_service(mocker):
 
 
 def test_get_final_image(mocker):
-    from derex.runner.docker import image_exists
+    from derex.runner.docker_utils import image_exists
 
     mocker.patch(
         "derex.runner.docker.docker.APIClient",
